Convert photos, events, tasks & badges lists to RecyclerViews w/ cached reusable cells & XML layouts; Support photo pinch-to-zoom, double-tap zoom; Fixed remaining tasks not showing in profile Progress tab; Convert about fragment to activity; Update Glide & Firebase Messaging to latest versions; Remove some unused resources

+++ isegoria/src/main/java/com/eulersbridge/isegoria/feed/PhotosFragment.java
@@ -156,7 +156,7 @@ public class PhotosFragment extends Fragment {
 	        textViewArticle.setOnClickListener(view12 -> {
 				PhotoAlbumFragment albumFragment = new PhotoAlbumFragment();
 				Bundle args = new Bundle();
-				args.putInt("albumId", albumId);
+				args.putLong("albumId", albumId);
 				albumFragment.setArguments(args);
 
 				getActivity().getSupportFragmentManager()
@@ -169,7 +169,7 @@ public class PhotosFragment extends Fragment {
 	       view.setOnClickListener(view1 -> {
 			   PhotoAlbumFragment albumFragment = new PhotoAlbumFragment();
 			   Bundle args = new Bundle();
-			   args.putInt("albumId", albumId);
+			   args.putLong("albumId", albumId);
 			   albumFragment.setArguments(args);
 
 			   getActivity().getSupportFragmentManager()
